// Client/src/app/grant-call/grant-call-interfaces.ts
export interface ResearchType {
    researchTypeCode: string;
    description: string;
    updateTimeStamp: number;
    updateUser: string;
    isActive: boolean;
}

export interface ResearchType2 {
    researchTypeCode: string;
    description: string;
    updateTimeStamp: number;
    updateUser: string;
    isActive: boolean;
}

export interface ResearchTypeArea {
    researchTypeAreaCode: string;
    description: string;
    researchTypeCode: string;
    researchType: ResearchType2;
    isActive: boolean;
    updateTimeStamp: number;
    updateUser: string;
}

export interface ResearchType3 {
    researchTypeCode: string;
    description: string;
    updateTimeStamp: number;
    updateUser: string;
    isActive: boolean;
}

export interface ResearchTypeArea2 {
    researchTypeAreaCode: string;
    description: string;
    researchTypeCode: string;
    researchType: ResearchType3;
    isActive: boolean;
    updateTimeStamp: number;
    updateUser: string;
}

export interface ResearchTypeSubArea {
    researchTypeSubAreaCode: string;
    description: string;
    researchTypeAreaCode: string;
    researchTypeArea: ResearchTypeArea2;
    isActive: boolean;
    updateTimeStamp: number;
    updateUser: string;
}

export interface areaOfResearch {
    grantResearchAreaId: number;
    grantCallId: number;
    researchTypeCode: string;
    researchType: ResearchType;
    researchTypeAreaCode: string;
    researchTypeArea: ResearchTypeArea;
    researchTypeSubAreaCode: string;
    researchTypeSubArea: ResearchTypeSubArea;
    updateTimeStamp: number;
    updateUser: string;
}